recall-etl-func/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
azure-storage-blob
python-dotenv
azure-core
azure-search-documents
openai
azure-identity
aiohttp
azure-data-tables